--- Projects/Team1/Prystopchuk/CCTask1_OOP/CCTask1.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30907.101
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CCTask1", "CCTask1\CCTask1.csproj", "{26828053-ADC6-4DEF-94BB-315D85ABB035}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{26828053-ADC6-4DEF-94BB-315D85ABB035}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{26828053-ADC6-4DEF-94BB-315D85ABB035}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{26828053-ADC6-4DEF-94BB-315D85ABB035}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{26828053-ADC6-4DEF-94BB-315D85ABB035}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {CEF2CEB5-CF96-4909-BEF5-7597C96C095C}
-	EndGlobalSection
-EndGlobal
