--- Projects/Team1/Prystopchuk/OOP_star/OOP_star.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30907.101
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OOP_star", "OOP_star\OOP_star.csproj", "{421C0B9C-48DA-41C0-A4C1-9C34423607B5}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{421C0B9C-48DA-41C0-A4C1-9C34423607B5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{421C0B9C-48DA-41C0-A4C1-9C34423607B5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{421C0B9C-48DA-41C0-A4C1-9C34423607B5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{421C0B9C-48DA-41C0-A4C1-9C34423607B5}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {E57F0E85-98F4-4BCA-8526-2255040CF413}
-	EndGlobalSection
-EndGlobal
