+++ Projects/Team1/Prystopchuk/OOP_star/Test.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace OOP_star
+{
+    public class Test
+    {
+
+    }
+}
